+++ src/Global-Components/button-add-a-card/ButtonAddCard.tsx
@@ -1,9 +1,9 @@
 import { useState } from "react";
 import styles from "./styles.module.css";
-import { postDataFromApi } from "../../../utils/functions-fetch";
+import { postDataFromApi } from "../../utils/functions-fetch";
 import { useNavigate, useParams } from "react-router-dom";
-import { DataListTask } from "../Board";
-import useUpdatePage from "../custom-hook/useUpdatePage";
+import { DataListTask } from "../../body/board/Board";
+import useUpdatePage from "../../custom-hook/useUpdatePage";
 
 type PropsTask = {
   task: DataListTask;
