+++ src/Global-Components/menu-drop-down/MenuDropDown.tsx
@@ -1,8 +1,8 @@
 import styles from "./styles.module.css";
 import classNames from "classnames";
-import more from "../../../assets/more.svg";
+import more from "../../assets/more.svg";
 import React, { useState } from "react";
-import useClickOutside from "../custom-hook/useClickOutSide";
+import useClickOutside from "../../custom-hook/useClickOutSide";
 
 type PropsBoard = {
   title: string;
