+++ src/components/body/header/Header.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from "react-router-dom";
-import logo from "../../assets/logo.svg";
+import logo from "../../../assets/logo.svg";
 import styles from "./styles.module.css";
 
 function Header() {
